# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "记忆 API",
        "title": "Remembrance API",
        "contact": {
            "name": "lzd",
            "email": "2804366305@qq.com"
        },
        "version": "1.0"
    },
    "host": "8.138.81.141",
    "basePath": "api",
    "paths": {
...
        "/api/common/photo/delete": {
            "post": {
                "description": "需要图片 photoid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "删除共同记忆",
                "parameters": [
                    {
                        "description": "photoid",
                        "name": "photoid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/common/photo/get": {
            "post": {
                "description": "需要 userid(用于记录搜索历史) location 传回的信息中包含url photoid text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "获取指定地点的共同记忆",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "location",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/common/photo/getself": {
            "post": {
                "description": "需要 userid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "获取自己发布的共同记忆",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
...
        "/api/photo/common/comment/get": {
            "get": {
                "description": "需要 photoid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "获取共同评论",
                "parameters": [
                    {
                        "description": "photoid",
                        "name": "photoid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/common/comment/getsearch": {
            "get": {
                "description": "需要 userid 返回20条搜素历史",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "获取搜素历史",
                "parameters": [
                    {
                        "description": "location",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/common/comment/post": {
            "put": {
                "description": "需要 UserId photoid text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "发布共同评论",
                "parameters": [
                    {
                        "description": "photoid",
                        "name": "photoid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "text",
                        "name": "text",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "groupid",
                        "name": "groupid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/common/comment/randget": {
            "get": {
                "description": "发送请求即可",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "获取随机共同记忆",
                "parameters": [
                    {
                        "description": "location",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/common/photo/post": {
            "put": {
                "description": "需要 UserId 图片url text location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "发布共同记忆",
                "parameters": [
                    {
                        "description": "cloudurl",
                        "name": "cloudurl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "text",
                        "name": "text",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "location",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/gettoken": {
            "get": {
                "description": "发送请求即可 发送请求即可 用于上传图片",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "获取qntoken 用于上传图片",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/group/post": {
            "put": {
                "description": "需要 UserId groupid 图片url text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "发布多人记忆",
                "parameters": [
                    {
                        "description": "cloudurl",
                        "name": "cloudurl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "text",
                        "name": "text",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "groupid",
                        "name": "groupid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/personal/createalbum": {
            "put": {
                "description": "获取 UserId  AlbumName",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "创建个人相册",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "personalalbumname",
                        "name": "personalalbumname",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/photo/personal/delete": {
            "post": {
                "description": "需要 UserId 相册id: personalalbumid  图片id：photoid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "controller"
                ],
                "summary": "删除个人记忆",
                "parameters": [
                    {
                        "description": "photoid",
                        "name": "photoid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "personalalbumid",
                        "name": "personalalbumid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    },
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.Message": {
            "type": "object",
            "properties": {
                "cloudurl": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "groupcode": {
                    "type": "string"
                },
                "groupid": {
                    "type": "integer"
                },
                "groupname": {
                    "type": "string"
                },
                "ifkeepgroupphoto": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "personalAlbumId": {
                    "type": "integer"
                },
                "personalalbumname": {
                    "type": "string"
                },
                "photoid": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
...
        "response.FailMesData": {
            "type": "object"
        },
        "response.OkMesData": {
            "type": "object"
        }
    }
}